+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved signed transaction handling to sui/sui_rpc.py from faux_wallet
 - Moved configuration management to ClientConfig and SuiConfig from faux_wallet
 - Moved content of sui/sui_tx_return_types to sui/sui_types
-- Fixed type in transaction return type class `Status`
+- Fixed typo in transaction return type class `Status`
 
 ### Removed
 - sui/sui_tx_return_types.py